Add near api update page and wallet collection
