+++ package.json
@@ -131,6 +131,7 @@
     "next-mdx-remote": "3",
     "next-sitemap": "^1.4.17",
     "node-cron": "^3.0.0",
+    "node-fetch": "^3.2.10",
     "node-sass": "^7.0.1",
     "nodemailer": "^6.6.3",
     "passport": "^0.4.1",
