+++ src/lib/routes.ts
@@ -47,6 +47,12 @@ export const routes = {
     },
     stepexample: {
       href: "/routes/kelvin/step_example"
+    },
+    nearupdate: {
+      href: "/routes/kelvin/near_api_update"
+    },
+    neartest: {
+      href: "/routes/kelvin/near_api_test"
     }
   }
 };
