+++ src/models/index.server.ts
@@ -16,8 +16,9 @@ import { Process } from "./process.server";
 import { Org } from "./org.server";
 import { StepInstance } from "./step_instance.server";
 import { Step } from "./step.server";
+import { Wallet } from "./wallet.server";
 
-const models = [User, SampleModel, Content, Resolution, Prediction, Value, Chat, Goal, Training, PastScore, Proj, ProcessTemplate, Process, Org, StepInstance, Step];
+const models = [User, SampleModel, Content, Resolution, Prediction, Value, Chat, Goal, Training, PastScore, Proj, ProcessTemplate, Process, Org, StepInstance, Step, Wallet];
